refactor: rename GroupPageLoader file
