+++ src/pages/Group/index.ts
@@ -6,5 +6,5 @@ export * from "./GroupIntro/GroupIntro"
 export * from "./GroupTabs"
 export * from "./GroupTabs/GroupTabs"
 export * from "./GroupContext"
-export * from "./loader"
+export * from "./GroupPageLoader"
 export * from "./GroupConfig/GroupConfigModal"
